+++ packages/sdk/typescript/src/Client.ts
@@ -8,6 +8,7 @@ import { DatasetsClient } from "./api/resources/datasets/client/Client.js";
 import { ExperimentsClient } from "./api/resources/experiments/client/Client.js";
 import { IncidentsClient } from "./api/resources/incidents/client/Client.js";
 import { MembersClient } from "./api/resources/members/client/Client.js";
+import { MemoryClient } from "./api/resources/memory/client/Client.js";
 import { MonitorsClient } from "./api/resources/monitors/client/Client.js";
 import { OauthKeysClient } from "./api/resources/oauthKeys/client/Client.js";
 import { ProjectsClient } from "./api/resources/projects/client/Client.js";
@@ -50,6 +51,7 @@ export class LatitudeClient {
     protected _spans: SpansClient | undefined;
     protected _experiments: ExperimentsClient | undefined;
     protected _sessions: SessionsClient | undefined;
+    protected _memory: MemoryClient | undefined;
 
     constructor(options: LatitudeClient.Options = {}) {
         this._options = normalizeClientOptionsWithAuth(options);
@@ -129,6 +131,10 @@ export class LatitudeClient {
 
     public get sessions(): SessionsClient {
         return (this._sessions ??= new SessionsClient(this._options));
+    }
+
+    public get memory(): MemoryClient {
+        return (this._memory ??= new MemoryClient(this._options));
     }
 
     /**